requirements.txt
fastapi
uvicorn[standard]
pydantic
beautifulsoup4
httpx
whois
dnspython